Allow Mallard links to be provided in an AsciiDoc attribute
Info links are an integral part of how Mallard works. This allows you
to specify info links as a space-separated list of tokens using the
'mallard-links' attribute in AsciiDoc. Each token specifies one link,
and is of the general form:

  type/group:xref

This produces:

  <link type="type" group="group" xref="xref"/>

Groups are optional. If no type is specified, 'guide' is assumed.